+++ test/test_context.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Unit tests for cartoframes.layers"""
+"""Unit tests for cartoframes.context"""
 import unittest
 import os
 import sys
@@ -13,7 +13,6 @@ import cartoframes
 from carto.exceptions import CartoException
 from carto.auth import APIKeyAuthClient
 from carto.sql import SQLClient
-from pyrestcli.exceptions import NotFoundException
 import pandas as pd
 import IPython
 from cartoframes.utils import dict_items
@@ -29,7 +28,7 @@ class _UserUrlLoader:
             try:
                 creds = json.loads(open('test/secret.json').read())
                 user_url = creds['USERURL']
-            except:
+            except:  # noqa: E722
                 warnings.warn('secret.json not found')
 
         if user_url in (None, ''):
@@ -77,41 +76,54 @@ class TestCartoContext(unittest.TestCase, _UserUrlLoader):
 
         # table naming info
         has_mpl = 'mpl' if os.environ.get('MPLBACKEND') else 'nonmpl'
+        has_gpd = 'gpd' if os.environ.get('USE_GEOPANDAS') else 'nongpd'
         pyver = sys.version[0:3].replace('.', '_')
+        buildnum = os.environ.get('TRAVIS_BUILD_NUMBER')
+
+        test_slug = '{ver}_{num}_{mpl}_{gpd}'.format(
+            ver=pyver, num=buildnum, mpl=has_mpl, gpd=has_gpd
+        )
 
         # test tables
         self.test_read_table = 'cb_2013_us_csa_500k'
         self.valid_columns = set(['affgeoid', 'aland', 'awater', 'created_at',
                                   'csafp', 'geoid', 'lsad', 'name', 'the_geom',
                                   'updated_at'])
-        table_args = dict(ver=pyver, mpl=has_mpl)
+        table_args = dict(ver=pyver, mpl=has_mpl, gpd=has_gpd)
         # torque table
         self.test_point_table = 'tweets_obama'
 
         # for writing to carto
-        self.test_write_table = 'cartoframes_test_table_{ver}_{mpl}'.format(
-            **table_args)
-        self.mixed_case_table = 'AbCdEfG_{0}_{1}'.format(pyver, has_mpl)
+        self.test_write_table = (
+            'cf_test_table_{}'
+        ).format(test_slug)
+
+        self.mixed_case_table = (
+            'AbCdEfG_{}'
+        ).format(test_slug)
 
         # for batch writing to carto
         self.test_write_batch_table = (
-            'cartoframes_test_batch_table_{ver}_{mpl}'.format(
-                **table_args))
+            'cf_testbatch_table_{}'
+        ).format(test_slug)
 
         self.test_write_lnglat_table = (
-            'cartoframes_test_write_lnglat_table_{ver}_{mpl}'.format(
-                **table_args))
+            'cf_testwrite_lnglat_table_{}'
+        ).format(test_slug)
 
         self.write_named_index = (
-                'cartoframes_test_write_non_default_index_{ver}_{mpl}'.format(
-                    **table_args))
+            'cf_testwrite_non_default_index_{}'
+        ).format(test_slug)
+
         # for queries
-        self.test_query_table = ('cartoframes_test_query_'
-                                 'table_{ver}_{mpl}'.format(
-                                    **table_args))
-        self.test_delete_table = ('cartoframes_test_delete_'
-                                  'table_{ver}_{mpl}').format(
-                                      **table_args)
+        self.test_query_table = (
+            'cf_testquery_table_{}'
+        ).format(test_slug)
+
+        self.test_delete_table = (
+            'cf_testdelete_table_{}'
+        ).format(test_slug)
+
         # for data observatory
         self.test_data_table = 'carto_usa_offices'
 
@@ -153,7 +165,8 @@ class TestCartoContext(unittest.TestCase, _UserUrlLoader):
     @unittest.skipIf(WILL_SKIP, 'no carto credentials, skipping this test')
     def test_cartocontext_credentials(self):
         """context.CartoContext.__init__ Credentials argument"""
-        creds = cartoframes.Credentials(username=self.username,
+        creds = cartoframes.Credentials(base_url=self.baseurl,
+                                        username=self.username,
                                         key=self.apikey)
         cc = cartoframes.CartoContext(creds=creds)
         self.assertIsInstance(cc, cartoframes.CartoContext)
@@ -161,17 +174,29 @@ class TestCartoContext(unittest.TestCase, _UserUrlLoader):
         self.assertEqual(cc.creds.key(), self.apikey)
 
         # CartoContext pulls from saved credentials
-        saved_creds = cartoframes.Credentials(username=self.username,
+        saved_creds = cartoframes.Credentials(base_url=self.baseurl,
+                                              username=self.username,
                                               key=self.apikey)
         saved_creds.save()
         cc_saved = cartoframes.CartoContext()
         self.assertEqual(cc_saved.creds.key(), self.apikey)
 
     @unittest.skipIf(WILL_SKIP, 'no carto credentials, skipping this test')
+    def test_cartocontext_authenticated(self):
+        """context.CartoContext._is_authenticated"""
+        with self.assertRaises(ValueError):
+            cc = cartoframes.CartoContext(
+                base_url=self.baseurl.replace('https', 'http'),
+                api_key=self.apikey
+            )
+
+    @unittest.skipIf(WILL_SKIP, 'no carto credentials, skipping this test')
     def test_cartocontext_isorguser(self):
         """context.CartoContext._is_org_user"""
-        cc = cartoframes.CartoContext(base_url=self.baseurl,
-                                      api_key=self.apikey)
+        cc = cartoframes.CartoContext(
+            base_url=self.baseurl,
+            api_key=self.apikey
+        )
         self.assertTrue(not cc._is_org_user())
 
     @unittest.skipIf(WILL_SKIP, 'no carto credentials, skipping this test')
@@ -934,6 +959,14 @@ class TestCartoContext(unittest.TestCase, _UserUrlLoader):
         meta_cols = set(('geom_id', 'geom_tags', 'geom_type', ))
         self.assertTrue(meta_cols & set(boundary_meta.columns))
 
+        # boundary metadata with correct timespan
+        meta_2015 = cc.data_boundaries(timespan='2015')
+        self.assertTrue(meta_2015[meta_2015.valid_timespan].shape[0] > 0)
+
+        # test for no data with an incorrect or invalid timespan
+        meta_9999 = cc.data_boundaries(timespan='invalid_timespan')
+        self.assertTrue(meta_9999[meta_9999.valid_timespan].shape[0] == 0)
+
         # boundary metadata in a region
         regions = (
             self.test_read_table,
@@ -972,7 +1005,7 @@ class TestCartoContext(unittest.TestCase, _UserUrlLoader):
                 'us.census.tiger.state' in set(meta.geom_id),
                 tf
             )
- 
+
         with self.assertRaises(ValueError):
             cc.data_boundaries(region=[1, 2, 3])
 
@@ -1083,17 +1116,24 @@ class TestCartoContext(unittest.TestCase, _UserUrlLoader):
                                      keywords='education')
             cc.data(self.test_read_table, meta)
 
-
     def test_column_name_collision_do_enrichement(self):
-        dup_col = 'female_third_level_studies_rate_2011'
-        self.sql_client.send("""create table {table} as (
-                select cdb_latlng(40.4165,-3.70256) the_geom,
-                       1 {dup_col})""". \
-                             format(dup_col=dup_col,
-                                    table=self.test_write_table))
+        """context.CartoContext.data column collision"""
+        dup_col = 'female_third_level_studies_2011_by_female_pop'
         self.sql_client.send(
-            "select cdb_cartodbfytable('public', '{table}')". \
-                format(table=self.test_write_table))
+            """
+            create table {table} as (
+                select cdb_latlng(40.4165,-3.70256) the_geom,
+                       1 {dup_col})
+            """.format(
+                dup_col=dup_col,
+                table=self.test_write_table
+            )
+        )
+        self.sql_client.send(
+            "select cdb_cartodbfytable('public', '{table}')".format(
+                table=self.test_write_table
+            )
+        )
 
         cc = cartoframes.CartoContext(base_url=self.baseurl,
                                       api_key=self.apikey)
@@ -1106,6 +1146,18 @@ class TestCartoContext(unittest.TestCase, _UserUrlLoader):
         )
 
         self.assertIn('_' + dup_col, data.keys())
+
+    def test_tables(self):
+        """context.CartoContext.tables normal usage"""
+        cc = cartoframes.CartoContext(
+            base_url=self.baseurl,
+            api_key=self.apikey
+        )
+        tables = cc.tables()
+        self.assertIsInstance(tables, list)
+        self.assertIsInstance(tables[0], cartoframes.analysis.Table)
+        self.assertIsNotNone(tables[0].name)
+        self.assertIsInstance(tables[0].name, str)
 
 
 class TestBatchJobStatus(unittest.TestCase, _UserUrlLoader):
@@ -1141,12 +1193,16 @@ class TestBatchJobStatus(unittest.TestCase, _UserUrlLoader):
         # sets skip value
         WILL_SKIP = self.apikey is None or self.username is None  # noqa: F841
         has_mpl = 'mpl' if os.environ.get('MPLBACKEND') else 'nonmpl'
+        has_gpd = 'gpd' if os.environ.get('HAS_GEOPANDAS') else 'nongpd'
+        buildnum = os.environ.get('TRAVIS_BUILD_NUMBER')
         pyver = sys.version[0:3].replace('.', '_')
 
         # for writing to carto
         self.test_write_lnglat_table = (
-            'cartoframes_test_write_lnglat_table_{ver}_{mpl}'.format(
+            'cf_test_write_lnglat_table_{ver}_{num}_{mpl}_{gpd}'.format(
                 ver=pyver,
+                num=buildnum,
+                gpd=has_gpd,
                 mpl=has_mpl))
 
     def tearDown(self):
@@ -1222,4 +1278,3 @@ class TestBatchJobStatus(unittest.TestCase, _UserUrlLoader):
         str_bjs = BatchJobStatus(cc, 'foo')
         self.assertIsNone(str_bjs.get_status())
         self.assertEqual(str_bjs.job_id, 'foo')
-